feat(mocks): consolidate and refactor mock data for quizzes and user quiz answers
- Moved quiz and user quiz answer mock data to separate files for better organization and maintainability.
- Updated references in syncUnansweredQuizIdsMocks.ts to utilize the new mock files, improving clarity in test setups.
- Removed redundant mock data from mockData.ts to streamline the codebase.
